scripts/pyairbyte/requirements.txt
airbyte
google-cloud-bigquery[pandas]
